First iteration: Linux compilation
