--- GMPHD_MAF/GMPHD_MAF.vcxproj.user
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <LocalDebuggerEnvironment>PATH=E:\OpenCV3.4.1\build_vc15_cuda1001\install\x64\vc15\bin;E:\boost_1_74_0\lib64-msvc-14.1;%PATH% $(LocalDebuggerEnvironment)</LocalDebuggerEnvironment>
-    <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
-  </PropertyGroup>
-</Project>
