+++ GMPHD_MAF/OnlineTracker.h
@@ -32,9 +32,10 @@ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#pragma once
+#ifndef ONLINE_TRACKER_H
+#define ONLINE_TRACKER_H
 
-#include "utils.hpp"
+#include "types.hpp"
 #include "hungarian.h"
 
 class OnlineTracker
@@ -155,3 +156,4 @@ protected:
 	}
 };
 
+#endif // ONLINE_TRACKER_H
