+++ GMPHD_MAF/hungarian.cpp
@@ -12,7 +12,6 @@
 // Both this code and the orignal code are published under the BSD license.
 // by Cong Ma, 2016
 // 
-#include "pch.h"
 #include "hungarian.h"
